langchain/retrievers/requirements.txt
langchain==0.3.4
langchain-chroma==0.1.4
langchain-community==0.3.3
langchain-huggingface==0.1.2
langchain-openai==0.2.3
lark==1.2.2
python-dotenv==1.0.1